feat: #10 🎸 nuxt3用にファイル名を修正
